packages/msaclean/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "msaclean"
version = "0.1.0"
description = "Clean and curate multiple sequence alignments for phylogenetic analysis"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "KamikiRitsu", email = "kamikiritsu510@gmail.com" }]
requires-python = ">=3.8"
dependencies = [
    "biopython>=1.80",
]

[project.urls]
Homepage = "https://github.com/KamikiRitsu510/msaclean"

[project.scripts]
msaclean = "msaclean.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
